Remove pytest-dependency
